+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **Auditing Changes**
   - Full scale replacement of sheet auditing module
+- **Design Changes**
+  - Mass importing sheets now notifies the user doing the import instead of the user who created the design
 - **Gem Changes**
   - Updated to rails 4.2.0.beta1
   - Updated to contour 2.6.0.beta4